Fix ValueError if a DOCSIS channel goes down.  Add example HTML for testing.
The modem changes all the fields from "Power" on out in the downstream table
to "----" if a channel is not getting sufficient signal to operate. Previous
code would die with a ValueError when it tried to turn it in to a number. Now
testing for this and returning None/null in all dashed out fields.

Added two example HTML dumps from my modem for testing.

cm820_8x2.html is "normal" operation with a low signal on three channels.
cm820_8x2_desynced_channel.html shows one of those going low enough to drop.

Use the command line option for the URL to point the script at these files.